+++ khoros/utils/tests/test_roles.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 """
-:Module:         khoros.utils.tests.test_roles
-:Synopsis:       This module is used by pytest to verify the :py:mod:`khoros.objects.roles` functionality.
-:Created By:     Jeff Shurtliff
-:Last Modified:  Jeff Shurtliff
-:Modified Date:  03 Oct 2022
+:Module:            khoros.utils.tests.test_roles
+:Synopsis:          This module is used by pytest to verify the :py:mod:`khoros.objects.roles` functionality.
+:Created By:        Jeff Shurtliff
+:Last Modified:     Jeff Shurtliff
+:Modified Date:     03 Oct 2022
 """
 
 import os
@@ -14,9 +14,6 @@ import pytest
 
 from . import resources
 from ...errors import exceptions
-
-# Define constants
-SKIP_LOCAL_TEST_MSG = 'skipping local-only tests'
 
 # Define a global variable to define when the package path has been set
 package_path_defined = False
@@ -39,6 +36,9 @@ def set_package_path():
 def test_get_role_id():
     """This function tests the :py:func:`khoros.objects.roles.get_role_id` function and corresponding method.
 
+    .. versionchanged:: 5.1.2
+       The function has been updated to support GitHub Workflows and to include a couple extra tests.
+
     .. versionadded:: 5.0.0
     """
     # Instantiate the core object
@@ -46,7 +46,17 @@ def test_get_role_id():
 
     # Test the method and function
     role_id = khoros_object.roles.get_role_id('Administrator')
-    assert role_id == 't:Administrator'          # nosec
+    assert role_id == 't:Administrator'                 # nosec
+    role_id = khoros_object.roles.get_role_id('some_role', 'board', 'some-blog-node')
+    assert role_id == 'b:some-blog-node:some_role'      # nosec
+
+    # Test that the correct exception is raised for an invalid scope
+    with pytest.raises(exceptions.InvalidRoleError):
+        khoros_object.roles.get_role_id('Administrator', 'something_wrong')
+
+    # Test that the correct exception is raised for a missing node ID
+    with pytest.raises(exceptions.MissingRequiredDataError):
+        khoros_object.roles.get_role_id('some_role', 'board')
 
 
 def test_invalid_role_type():
@@ -62,14 +72,13 @@ def test_invalid_role_type():
 def test_total_role_type_counts():
     """This function tests the :py:meth:`khoros.core.Khoros.Role.get_total_role_count` method and related function.
 
+    .. versionchanged:: 5.1.2
+       The function has been updated to support GitHub Workflows.
+
     .. versionadded:: 5.0.0
     """
-    if not resources.local_test_config_exists() or not resources.local_helper_exists():
-        pytest.skip(SKIP_LOCAL_TEST_MSG)
-
-    # Instantiate the Khoros object
-    set_package_path()
-    khoros_object = resources.instantiate_with_local_helper()
+    # Instantiate the core object
+    khoros_object = resources.get_core_object()
 
     # Ensure that the default result is an integer of the total count
     total_count = khoros_object.roles.get_total_role_count()
@@ -95,14 +104,13 @@ def test_total_role_type_counts():
 def test_get_roles_for_user():
     """This function tests the :py:meth:`khoros.core.Khoros.Role.get_roles_for_user` method and related function.
 
+    .. versionchanged:: 5.1.2
+       The function has been updated to support GitHub Workflows.
+
     .. versionadded:: 5.0.0
     """
-    if not resources.local_test_config_exists() or not resources.local_helper_exists():
-        pytest.skip(SKIP_LOCAL_TEST_MSG)
-
-    # Instantiate the Khoros object
-    set_package_path()
-    khoros_object = resources.instantiate_with_local_helper()
+    # Instantiate the core object
+    khoros_object = resources.get_core_object()
 
     # Test the method and function using an integer and a string as the User ID
     for user_id in [1, '1']:
@@ -116,14 +124,13 @@ def test_get_roles_for_user():
 def test_get_users_with_role():
     """This function tests the :py:meth:`khoros.core.Khoros.Role.get_users_with_role` method and related function.
 
+    .. versionchanged:: 5.1.2
+       The function has been updated to support GitHub Workflows.
+
     .. versionadded:: 5.0.0
     """
-    if not resources.local_test_config_exists() or not resources.local_helper_exists():
-        pytest.skip(SKIP_LOCAL_TEST_MSG)
-
-    # Instantiate the Khoros object
-    set_package_path()
-    khoros_object = resources.instantiate_with_local_helper()
+    # Instantiate the core object
+    khoros_object = resources.get_core_object()
 
     # Test the standard return mode
     users_with_role = khoros_object.roles.get_users_with_role(role_name='Administrator')
@@ -133,6 +140,6 @@ def test_get_users_with_role():
 
     # Test the simple return mode
     users_with_role = khoros_object.roles.get_users_with_role(role_name='Administrator', simple=True)
-    assert isinstance(users_with_role, list)          # nosec
+    assert isinstance(users_with_role, list)                # nosec
     if len(users_with_role) > 0:
         assert isinstance(users_with_role[0], str)          # nosec